Add RawTextHelpFormatter to install_parser
The `aqt install -h` command prints the help text for the `arch`
argument all together on one line. I noticed that the help text
is formatted very carefully in code, with lots of meaningful
whitespace, but on my computer, all that whitespace is being ignored.

With `formatter_class=argparse.RawTextHelpFormatter` added to
the `subparsers.add_parser()` function call, all of the whitespace
in the help text is preserved.

It looks like `formatter_class=argparse.RawTextHelpFormatter`
already exists in the call that creates the root parser, but it
appears that this property does not extend to all subparsers
automatically.