Add control and test for patched qmake
This adds two new jobs to the azure pipeline: one that tests a really
short output directory, and one for a really long output directory.

This also adds a QT_BINDIR variable, which is dependent on the output
directory. Within `steps.yml`, QT_BINDIR always points to the directory
that contains qmake.