Fix `test_cli_unknown_version`
The test test_cli_unknown_version, in `tests/test_connection.py`, is
never run in CI because it is marked as requiring remote data. When I
run `pytest` locally, this test always fails because the version string
"5.12" cannot be parsed into a valid `semantic_version.Version` object.
This change modifies that version string to "5.16.0", so that it will
make a valid Version object that will never be a valid version of Qt.

This also modifies the assertion code in test_cli_unknown_version.
On my machine, the new logger prints directly to `stderr` and not
`stdout` when run via `pytest`, and the `aqt - WARNING` and `aqt -ERROR`
prefixes are omitted. The new assertion code uses a regex to pick out
the required parts of the output and ignore anything that might not show
up.

This change also accounts for what happens during a redirect. In my
testing, I noticed that redirects would occasionally insert some
additional lines of output regarding a redirect. This change describes
this output in a comment, and swallows up the additional lines using a
Kleene star.