Fix multiple redirects in getUrl
This change handles multiple redirects. Without this change, if
`helper.altlink()` returns a url that will redirect, we cannot follow
that redirect. This change allows up to 10 such redirects to occur.

Please note that the upper bound for number of redirects was chosen
arbitrarily, with no regard for what might realistically occur. I felt
that it was important not to allow recursive or infinite redirects, and
that at least two redirects should be allowed. Perhaps the upper bound
should be modifiable in settings.ini.