Fix error message for non-200 status code
The original error message, `Download error when access <filename>`,
is confusing when a user is trying to use `aqt list`. This message
could make the user think that `aqt list` is attempting to download
archives, when in reality the command is only reading files on the
server.

Initially, I wanted to make this error message modifiable by the caller.
On further investigation, I realized that this function is always used
to retrieve 'Updates.xml' files or hashes, and that the error message
"Failed to retrieve file" would be appropriate for all usages.

Admittedly, the distinction between "Download error" and "Failed to
retrieve file" is subtle. However, the difference was big enough to
convince me as a developer that I should not use this function to
retrieve "Updates.xml" files, when in fact it was the right function to
use.