Refactor `suggested_follow_up`
This alters `suggested_follow_up` to return a list of strings.
This makes testing easier, as well as making it easier to change the
way the suggestions are displayed.

This change also fixes an error in the suggestions ('list' was missing
from the suggested command), and fixes some punctuation errors.

This also adds a print formatter that displays the suggestions nicely
in the terminal.