Simplify 'general usage' in installation.rst
This change removes most of the options listed in the 'general usage'
sections in `docs/installation.rst`, and formats them so they are easy
to read. In this section, users are looking for a simple description of
what the commands do, and how to use them. This change retains the most
commonly used options, and asks the reader to refer to `docs/cli.rst`
for more complete documentation.

This change also clarifies some of the descriptions of the most
commonly-used options.