Align cli.rst with actual interface
At some point, it looks like the file `docs/cli.rst` got misaligned with
the actual command line interface, and this change fixes that.

This moves the `version` subcommand into the 'generic commands' section.

This adds several command line options to `aqt install-qt` that were
missing from the docs, and removes `--version`, which is not an option
for `aqt install-qt`.