Fix use of version and version_dir
This changes `Updater.update()` to use two separate variables:
`version`, which is a Semantic Version, and `version_dir`, which is
a string for the directory in which Qt is installed.

`version_dir` cannot just be `str(version)`; in the special case of
Qt 5.9.0, `version_dir` must be "5.9" to account for the way that this
version of Qt was packaged.

(cherry picked from commit d6127eb8a00dfeeeedb0fd44db3e3529a115bb67)