Remove incorrect combinations checks
This combination-checking code is both obsolete and invalid.
All of these checks are done better by the `combinations.json` file,
which is now kept current by a bot running in CI.

The existing version check does not account for Qt 6.2, which leads to
an inappropriate warning when you install Qt 6.2 on windows desktop.