Allow test_doc_archives to run independently
When you run `test_doc_archives.test_parse_update_xml` outside the
test suite, the settings.ini file is not loaded. When you run it
together with all the other tests, it depends on the settings loaded
during another test.

This change loads the settings.ini file before each test in this file.