Add name of PackageUpdate to QtPackage
This adds the "Name" property from the "PackageUpdate" element from
"Updates.xml" to the QtPackage objects. This is intended to make
testing easier: if the "Name" property is missing, there's no way to
find out which PackageUpdate the 7z archive came from after it has
been added to the QtArchive.

In the case of ToolArchive contents, the `QtPackage.pkg_update_name`
is redundant, because the "Name" property is already saved as
`QtPackage.name`. The `QtPackage.name` property is being used in two
different ways in two different places, probably because of different
naming conventions.