Test SrcDocExamplesArchives properly
The existing tests in test_doc_archives.py don't actually test the
SrcDocExamplesArchives object; for some reason, they just test
QtArchives. The old tests make it hard to change the behavior of
QtArchives without causing these tests to fail.

This change alters these tests so that they use SrcDocExamplesArchives,
which is otherwise untested. It also changes the name of the "arch"
variable to "flavor", which is more accurate for the
SrcDocExamplesArchives class. It also adds tests for the untested
flavors, "src" and "examples".