Fix broken mirror link
This is meant to allow CI builds to run successfully when the mirror link in `cli.rst` is broken. It accomplishes this by changing the link into an html 'code' tag. The link is not meant to be clicked on; it is meant to be typed as a command line argument. This change makes the intention more explicit.

It may be better to replace the link with a working URL, perhaps one of the links in the settings.ini file.