Refactor parser creation
This reduces code duplication, with the intent of initializing the
parsers more consistently.

Why:
In #413, it was discovered that the interface of `aqt tool` did not
match the stated interface in the v1.2.5 documentation. After that,
I found that the interface for `aqt examples` and `aqt src` did not
fit that documentation either. This refactor places all of the
interface definitions in one location, with common code paths to
create each parser. This removes most of the code duplication involved
here, so that each parser is built more consistently. It also makes
it easier to visually inspect the differences between the parsers.