Update configuration.rst with new options
This updates the docs for the `settings.ini` file.

The purpose of this change is to deprecate `max_retries` and document
`max_retries_on_connection_error` and `max_retries_on_checksum_error`.
However, it appears that `docs/configuration.rst` has been out of date
for a while, since several entries are missing. This change adds all
the missing entries, along with the new entries.