Fix crash when target architecture was missing
This command would trigger a crash:
`aqt install-qt linux desktop 6.2.* --archives qtbase`
This is due to the `arch` argument was missing and it defaults to
`None`, however follow-up function calls expect it is a string and
results in a crash. This patch changed the default value to an empty
string.