Change `target` param for install-src/doc/example
This changes the CLI so that the `target` parameter for `install-src`,
`install-doc`, and `install-example` is now ignored, and "desktop" is
substituted for anything the user inputs. The only valid option here
is "desktop" anyway, so the user should be forgiven for trying to input
anything else.

It may be better to remove `target` from the parameter list, but that's
a big change to the interface that could break a lot of dependent code.