Improve CI & tests for install-src/doc/example
This updates CI and unit tests to monitor behavior of
`install-src/doc/example`, now that the `target` parameter has been
deprecated. This makes sure that the `target` parameter is optional,
and that the CLI warns the user that `target` is deprecated when a user
uses it.

[CI tests]:
This adds a test to `install-src/doc/example` Azure Pipeline jobs to
ensure that the desired files are installed, using `ls`.

This also changes some of the archives installed to `qtlottie`, which
is a particularly small file.

This also adds a build job for `install-example`.

[Unit tests]:
This adds tests for `install-src`, with and without the `target`
parameter.

It also fixes existing tests for `install-doc`, to account for
changes to the `target` parameter.