Improve Azure tests for install-src/doc/example
This change updates the usage of `install-src`, `install-doc`, and
`install-example` to omit the optional `TARGET` parameter.
It also changes the usage of `--archives`, so that the archives for
the modules are not be specified.

This also adds customizable checks to each build job, so that specific
requirements can be made on an individual basis.
The three `install-src/doc/example` commands now have specific checks
that will cause the build job to fail if the requested modules and
archives have not been installed.