cache powershell modules
This change saves PowerShell modules used in the Azure Pipelines to a
cache, and loads them when required. This is intended to prevent random
failures when the PowerShell Gallery fails to provide the requested
modules. These failures occur at random, possibly as a result of a
backpressure mechanism.

This change will also speed up all Windows Azure Pipelines jobs, since
downloads from the PowerShell Gallery will no longer be necessary.