Use Qt6 version of example project
The directory structure of the Qt libraries has changed enough that
the Qt5 example project we are using for WASM cannot be built using
Qt6. This change adds an example project appropriate for use with
Qt6, and adds build jobs for Mac and Linux that build it too,
to prove that we can build the same code on all 3 platforms.