Add 'Info.plist' file to accelbubble.7z
This `Info.plist` file is sourced from the Qt 6.4.0 version of the
`accelbubble` example project.

According to the qmake documentation, the Info.plist file should be
generated automatically by qmake. For whatever reason, this is not
happening on its own. The `accelbubble.pro` file specifies
`QMAKE_INFO_PLIST = Info.plist`, which (I think) is the code you would
use to specify an alternate Info.plist file. Perhaps qmake is not
generating the file because it thinks that you intended to read it from
the filesystem? If this is the case, then adding the Info.plist file
may solve the problem.