refactor: specify emsdk_tag directly
Previously, emsdk_tag was appended to emsdk_version, with an `@`
to delimit the two values. Additionally, a default tag was appended
after the specified tag. This makes the workflow unnecessarily
difficult to read and to debug, when compared to a more declarative
workflow style.

This work is being done when the default emsdk version and tag are
not working at all for several builds.