Fix patching for Qt 6.10.1 cross compile.
Qt 6.10.1 cross compile prepends "host-" to the qmake, qmake6,
qtpaths, qtpaths6 .bat files.  This has come up before in 6.10.0
betas, but not in the released 6.10.1.
To accomodate this variabilty in naming we check if the usual bat file
exists and patch that if it does, otherwise we attempt to patch a bat
file with "host-" prepended.